+++ apps/design-studio/schemaTypes/documents/article.tsx
@@ -1,5 +1,6 @@
 import { MdAddLink, MdArticle } from "react-icons/md";
 import { defineField, defineType } from "sanity";
+import { API_VERSION } from "../../sanity.config";
 
 // This is the ID of the components category
 // It's used to filter out a few fields when editing an article that's a
@@ -19,6 +20,20 @@ export const article = defineType({
       validation: (Rule) => Rule.required(),
     }),
     defineField({
+      name: "section",
+      title: "Section",
+      type: "reference",
+      to: [{ type: "section" }],
+      validation: (Rule) => Rule.required(),
+    }),
+    defineField({
+      name: "category",
+      title: "Category",
+      type: "reference",
+      to: [{ type: "category" }],
+      validation: (Rule) => Rule.required(),
+    }),
+    defineField({
       name: "slug",
       title: "Slug",
       type: "slug",
@@ -27,10 +42,38 @@ export const article = defineType({
       },
     }),
     defineField({
-      name: "category",
-      title: "Category",
-      type: "reference",
-      to: [{ type: "category" }],
+      name: "path",
+      title: "Path",
+      type: "slug",
+      options: {
+        slugify: (input: string) => {
+          return input.trim();
+        },
+        source: async (parent, context) => {
+          const sectionID = (parent.section as any)?._ref;
+          const sectionLabel = async () => {
+            const { getClient } = context;
+            const client = getClient({ apiVersion: API_VERSION });
+            const sectionDoc = await client.fetch(`*[_id == $sectionID]`, {
+              sectionID,
+            });
+            return sectionDoc?.[0].slug.current;
+          };
+          const categoryID = (parent.category as any)?._ref;
+          const categoryLabel = async () => {
+            const { getClient } = context;
+            const client = getClient({ apiVersion: API_VERSION });
+            const categoryDoc = await client.fetch(`*[_id == $categoryID]`, {
+              categoryID,
+            });
+            return categoryDoc?.[0].slug.current;
+          };
+          if (!sectionID) return "/";
+          return `${await sectionLabel()}/${await categoryLabel()}/${(parent.slug as { current: string }).current}`;
+        },
+      },
+      description:
+        "The full path to this page, click 'Generate' to update it when the title or section changes.",
       validation: (Rule) => Rule.required(),
     }),
     defineField({
@@ -101,6 +144,7 @@ export const article = defineType({
       type: "array",
       of: [
         { type: "articleHeader", title: "Lead paragraph" },
+        { type: "textBlock", title: "Text block" },
         { type: "textBlocks", title: "Text blocks" },
         { type: "imageBlock", title: "Image block" },
         { type: "imageAndTextList", title: "Image and text" },