+++ apps/design-studio/schemaTypes/documents/page.tsx
@@ -1,5 +1,6 @@
 import { MdAutoStories } from "react-icons/md";
 import { defineField, defineType } from "sanity";
+import { API_VERSION } from "../../sanity.config";
 
 export const page = defineType({
   name: "page",
@@ -33,6 +34,14 @@ export const page = defineType({
       group: "pageContent",
     }),
     defineField({
+      name: "section",
+      title: "Section",
+      type: "reference",
+      to: [{ type: "section" }],
+      validation: (Rule) => Rule.required(),
+      group: "pageContent",
+    }),
+    defineField({
       name: "slug",
       title: "Slug",
       type: "slug",
@@ -42,11 +51,38 @@ export const page = defineType({
       group: "pageContent",
     }),
     defineField({
+      name: "path",
+      title: "Path",
+      type: "slug",
+      options: {
+        slugify: (input: string) => {
+          return input.trim();
+        },
+        source: async (parent, context) => {
+          const sectionID = (parent.section as any)?._ref;
+          const sectionLabel = async () => {
+            const { getClient } = context;
+            const client = getClient({ apiVersion: API_VERSION });
+            const sectionDoc = await client.fetch(`*[_id == $sectionID]`, {
+              sectionID,
+            });
+            return sectionDoc?.[0].slug.current || "Unknown Section";
+          };
+          if (!sectionID) return "/";
+          return `${await sectionLabel()}/${(parent.slug as { current: string }).current}`;
+        },
+      },
+      description:
+        "The full path to this page, click 'Generate' to update it when the title or section changes.",
+      validation: (Rule) => Rule.required(),
+    }),
+    defineField({
       name: "content",
       title: "Content",
       type: "array",
       of: [
         { type: "articleHeader", title: "Lead paragraph" },
+        { type: "textBlock", title: "Text block" },
         { type: "textBlocks", title: "Text blocks" },
         { type: "imageBlock", title: "Image block" },
         { type: "imageAndTextList", title: "Image and text" },