+++ apps/designmanual-frontend/app/features/portable-text/PortableText.tsx
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+
 import { ButtonProps } from "@chakra-ui/react";
 import {
   PortableText as SanityPortableText,
@@ -18,7 +19,6 @@ import {
   Image,
   List,
   ListItem,
-  Separator,
   SimpleGrid,
   StaticCard,
   Text,
@@ -30,11 +30,23 @@ import { Link } from "react-router";
 
 import { urlBuilder } from "~/utils/sanity/utils";
 
-import { ComponentDocs } from "../../routes/_base.$category.$slug/component-docs/ComponentDocs";
+import { ComponentDocs } from "../../routes/_base.$section.$category.$slug/component-docs/ComponentDocs";
 import { CodeBlock } from "./code-block/CodeBlock";
 import { ImageWithCaption } from "./ImageWithCaption";
 import { InteractiveCode } from "./interactive-code/InteractiveCode";
 import { LinkableHeading } from "./LinkableHeading";
+import { AccordionSerializer } from "./serializers/AccordionSerializer";
+import { ArticleHeaderSerializer } from "./serializers/ArticleHeaderSerializer";
+import { CardSerializer } from "./serializers/CardSerializer";
+import { DividerSerializer } from "./serializers/DividerSerializer";
+import { FileListSerializer } from "./serializers/FileListSerializer";
+import { ImageAndTextListSerializer } from "./serializers/ImageAndTextListSerializer";
+import { ImageBlockSerializer } from "./serializers/ImageBlockSerializer";
+import { ImageCardListSerializer } from "./serializers/ImageCardListSerializer";
+import { LinkButtonSerializer } from "./serializers/LinkButtonSerializer";
+import { NonClickableBoxListSerializer } from "./serializers/NonClickableBoxesSerializer";
+import { TextBlockSerializer } from "./serializers/TextBlockSerializer";
+import { TextBlocksSerializer } from "./serializers/TextBlocksSerializer";
 
 const components: Partial<PortableTextReactComponents> = {
   marks: {
@@ -94,7 +106,7 @@ const components: Partial<PortableTextReactComponents> = {
       </Heading>
     ),
     normal: ({ children }) => {
-      const arrayChildren = React.Children.toArray(children);
+      const arrayChildren = React.Children.toArray(children as React.ReactNode);
       if (arrayChildren.length === 0 || arrayChildren.join("") === "") {
         return null;
       }
@@ -133,6 +145,18 @@ const components: Partial<PortableTextReactComponents> = {
     </ListItem>
   ),
   types: {
+    articleHeader: ArticleHeaderSerializer,
+    textBlock: TextBlockSerializer,
+    textBlocks: TextBlocksSerializer,
+    imageBlock: ImageBlockSerializer,
+    imageAndTextList: ImageAndTextListSerializer,
+    imageCardList: ImageCardListSerializer,
+    cards: CardSerializer,
+    linkButton: LinkButtonSerializer,
+    nonClickableBoxList: NonClickableBoxListSerializer,
+    accordion: AccordionSerializer,
+    fileList: FileListSerializer,
+    divider: DividerSerializer,
     buttonLink: ({ value }) => {
       const isInternal = value.url.startsWith("/");
       const linkProps = isInternal
@@ -146,7 +170,6 @@ const components: Partial<PortableTextReactComponents> = {
         </Box>
       );
     },
-    divider: () => <Separator height="1px" marginY={8} />,
     grid: ({ value }) => (
       <SimpleGrid
         columns={[1, 1, 2, value.maxNumberOfColumns]}