+++ apps/designmanual-frontend/app/routes/_base/content-menu/SearchableContentMenu.tsx
@@ -5,20 +5,24 @@ import { ContentMenu } from "./ContentMenu";
 
 export const SearchableContentMenu = () => {
   const focusableRef = React.useRef<HTMLButtonElement>(null);
+  const [refreshKey, setRefreshKey] = React.useState(0);
+
+  const handleRefresh = () => {
+    setRefreshKey((prev) => prev + 1);
+  };
 
   return (
     <Box
       as="nav"
-      aria-label={"Content"}
-      width={"100%"}
-      position="sticky"
-      top="7.15rem"
-      overflowY="scroll"
-      height="90vh"
-      paddingBottom={3}
-      paddingRight={3}
+      width={["90vw", null, null, "100%"]}
+      padding={0}
+      alignSelf={"flex-start"}
     >
-      <ContentMenu ref={focusableRef} />
+      <ContentMenu
+        ref={focusableRef}
+        refreshKey={refreshKey}
+        handleRefresh={handleRefresh}
+      />
     </Box>
   );
 };